refactor(uniswap): move chainIds to constants file
